Code cleanup
- Reformatted according to .editorconfig settings.

+++ src/Bach.Model.Test/BachXUnitSerializer.cs
@@ -1,20 +1,20 @@
 // Module Name: BachXUnitSerializer.cs
 // Project:     Bach.Model.Test
-// Copyright (c) 2012, 2025  Eddie Velasquez.
-//
+// Copyright (c) 2012, 2026  Eddie Velasquez.
+// 
 // This source is subject to the MIT License.
 // See http://opensource.org/licenses/MIT.
 // All other rights reserved.
-//
+// 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this software
 // and associated documentation files (the "Software"), to deal in the Software without restriction,
 // including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense,
 // and/or sell copies of the Software, and to permit persons to whom the Software is furnished to
 // do so, subject to the following conditions:
-//
+// 
 // The above copyright notice and this permission notice shall be included in all copies or substantial
 // portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 // INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
 // PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
@@ -22,6 +22,7 @@
 // CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
 // OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+using System.Diagnostics.CodeAnalysis;
 using Bach.Model;
 using Bach.Model.Test;
 using Xunit.Sdk;
@@ -40,9 +41,6 @@ using Xunit.Sdk;
 
 namespace Bach.Model.Test;
 
-using System.Diagnostics.CodeAnalysis;
-using Xunit.Sdk;
-
 internal class BachXUnitSerializer: IXunitSerializer
 {
   #region Public Methods
@@ -53,14 +51,14 @@ internal class BachXUnitSerializer: IXunitSerializer
   {
     return type switch
     {
-      not null when type == typeof( Accidental )      => (Accidental) int.Parse( serializedValue ),
-      not null when type == typeof( NoteName )        => (NoteName) int.Parse( serializedValue ),
-      not null when type == typeof( Pitch )           => (Pitch) int.Parse( serializedValue ),
-      not null when type == typeof( PitchClass )      => (PitchClass) int.Parse( serializedValue ),
-      not null when type == typeof( Interval )        => (Interval) int.Parse( serializedValue ),
-      not null when type == typeof( ModeFormula )     => (ModeFormula) int.Parse( serializedValue ),
-      not null when type == typeof( ScaleDegree )     => ScaleDegree.ScaleDegrees[int.Parse( serializedValue ) - 1],
-      _                                               => throw new NotSupportedException( $"Type {type} is not supported." )
+      not null when type == typeof( Accidental )  => (Accidental) int.Parse( serializedValue ),
+      not null when type == typeof( NoteName )    => (NoteName) int.Parse( serializedValue ),
+      not null when type == typeof( Pitch )       => (Pitch) int.Parse( serializedValue ),
+      not null when type == typeof( PitchClass )  => (PitchClass) int.Parse( serializedValue ),
+      not null when type == typeof( Interval )    => (Interval) int.Parse( serializedValue ),
+      not null when type == typeof( ModeFormula ) => (ModeFormula) int.Parse( serializedValue ),
+      not null when type == typeof( ScaleDegree ) => ScaleDegree.ScaleDegrees[int.Parse( serializedValue ) - 1],
+      _                                           => throw new NotSupportedException( $"Type {type} is not supported." )
     };
   }
 
@@ -85,14 +83,14 @@ internal class BachXUnitSerializer: IXunitSerializer
   {
     return value switch
     {
-      Accidental accidental           => ( (int) accidental ).ToString(),
-      NoteName noteName               => ( (int) noteName ).ToString(),
-      Pitch pitch                     => ( (int) pitch ).ToString(),
-      PitchClass pitchClass           => ( (int) pitchClass ).ToString(),
-      Interval interval               => ( (int) interval ).ToString(),
-      ModeFormula modeFormula         => ( (int) modeFormula ).ToString(),
-      ScaleDegree scaleDegree         => ( (int) scaleDegree.Degree ).ToString(),
-      _                               => throw new NotSupportedException( $"Type {value.GetType()} is not supported." )
+      Accidental accidental   => ( (int) accidental ).ToString(),
+      NoteName noteName       => ( (int) noteName ).ToString(),
+      Pitch pitch             => ( (int) pitch ).ToString(),
+      PitchClass pitchClass   => ( (int) pitchClass ).ToString(),
+      Interval interval       => ( (int) interval ).ToString(),
+      ModeFormula modeFormula => ( (int) modeFormula ).ToString(),
+      ScaleDegree scaleDegree => scaleDegree.Degree.ToString(),
+      _                       => throw new NotSupportedException( $"Type {value.GetType()} is not supported." )
     };
   }
